% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{as.data_model.data.frame}
\alias{as.data_model.data.frame}
\title{as.data_model.data.frame() is taken from bergant's datamodelr v0.2.1. Unfortunately it's not publically exposed and I need it so I'm importing it. All credit to bergant for this
Coerce a data frame to a data model}
\usage{
\method{as.data_model}{data.frame}(x)
}
\description{
as.data_model.data.frame() is taken from bergant's datamodelr v0.2.1. Unfortunately it's not publically exposed and I need it so I'm importing it. All credit to bergant for this
Coerce a data frame to a data model
}
