% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{dbplyr_to_sql}
\alias{dbplyr_to_sql}
\title{Tibble to T-SQL}
\usage{
dbplyr_to_sql(TibbleDbPointer, con = db$con)
}
\arguments{
\item{TibbleDbPointer}{A tbl()}

\item{con}{is a dbConnect {DBI} connection object to a SQL Database}
}
\description{
Retrieves the T-SQL code behind any Tbl even after R transformations
}
\examples{
dbplyr_to_sql(tbl("TableName", db$con) db$con)
}
\keyword{SQL}
\keyword{TSQL}
