% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{edit_db_fields}
\alias{edit_db_fields}
\title{Edits a db_fields formatted DF}
\usage{
edit_db_fields(db_fields)
}
\arguments{
\item{db_fields}{A DF. A dataframe formatted according to the db_fields standard.}
}
\description{
Opens a browser window where the user can select which columns to include and see the relationships amongst the tables.
When the browser window closes, the results are saved on the DF, but NOT on the db_fields.csv file; for this use write_db_fields_csv()
}
\examples{
db_fields <- read_db_fields_csv("db_fields.csv")
edit_db_fields(db_fields)
write_db_fields_csv(db_fields, "db_fields.csv")
}
\keyword{db_fields}
\keyword{dbfields}
\keyword{edit}
