% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{read_db_fields_csv}
\alias{read_db_fields_csv}
\title{Reads the csv file that is formatted according to the db_fields standard}
\usage{
read_db_fields_csv(
  csv_path,
  Update_DBEnv_DBFields = FALSE,
  UpdateDBFieldsFromDBCon = FALSE,
  ExcludeIdentities = FALSE,
  ExcludeForeignKeys = TRUE,
  ExcludeAuditingFields = FALSE,
  ExcludeSYSDIAGRAMS = TRUE
)
}
\arguments{
\item{csv_path}{A String. The path of the db_fields .csv file.}

\item{Update_DBEnv_DBFields}{A Boolean. If set to TRUE, you can have an internal main db_fields accessible via db$db_fields usually acting as the main db_fields. Default is FALSE as having local db_fields variables is the default behaviour.}

\item{UpdateDBFieldsFromDBCon}{A Boolean. If set to TRUE, then the db_fields saved on the file will be updated to reflect any changes in the SQL Database (removal or addition of Columns and Tables). User options are kept.}

\item{ExcludeIdentities}{A Boolean. The default inclusion behaviour for SQL Table Identities. If we're going to perform more joins afterwards or if we need to have a reference as to which table each column belongs to, then we shouldn't exclude them; if we only care about extrapolating information from each specific row, then we can exclude them by default and edit to include just the row identifier ID.}

\item{ExcludeForeignKeys}{A Boolean. If we need to perform manual joins afterwards or keep a reference as to with which table there's a connection, then we need them; otherwise we can safely exclude them}

\item{ExcludeAuditingFields.}{A Boolean. If TRUE, any SQL Columns ending with "_OrigEntryOn", "_OrigEntryBy", "_EntryOn", "_EntryBy", "_CompName", "_Remote" or "_Username" will have INCLUDE == FALSE by default}

\item{ExcludeSYSDIAGRAMS.}{A Boolean. If TRUE, any SQL Columns on the table "sysdiagrams" will have INCLUDE == FALSE by default}
}
\description{
Returns the db_fields, a DF with columns: KeyType, Table, Column, Type, Comment, RelationshipWithTable, RelationshipWithColumn, Transformation
}
\examples{
db_fields <- read_db_fields_csv("db_fields.csv")
}
\keyword{Read}
\keyword{db_fields}
\keyword{dbfields}
