% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{show_ER_diagramme}
\alias{show_ER_diagramme}
\title{Plots the ER Digramme for the SQL Database}
\usage{
show_ER_diagramme(DataModel, GraphDirection = "BT")
}
\arguments{
\item{DataModel}{A data model object as given by the datamodelr Library/Package}

\item{GraphDirection}{A string. Valid values are given by the rankdir parameter of the datamodelr Library/Package}
}
\description{
Shows a plot of the ER diagramme for the Database
}
\examples{
show_ER_diagramme(db$dm_f)
}
\keyword{Diagram}
\keyword{Diagramme}
\keyword{ER}
