% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{update_db_info}
\alias{update_db_info}
\title{Creates the db Environment and the SQL Connection (db$con)}
\usage{
update_db_info(Driver, Database, Server, UID, PWD, Trusted_Connection,
  Port = 1433, ...)
}
\arguments{
\item{Driver}{A string. A SQL Driver, etc.: {SQL Server};}

\item{Database}{A string. The name of the SQL Database itself}

\item{Server}{A string. A name that resolves into the SQL Machine, or the IP to the SQL Machine}

\item{UID}{A string. The Username credential for the SQL Connection. NULL if Windows Login is to be used}

\item{PWD}{A string. The Password credential for the SQL Connection. NULL if Windows Login is to be used}

\item{Trusted_Connection}{A Boolean. TRUE or FALSE on whether Windows Login is to be used}

\item{Port}{An Integer. The Port which the SQL Listener is listening at}

\item{...}{If the relationships exist on the SQL Database, this can be left blank. Logical expressions for the SQL Relationships in the format: table1$FKcolumn2 == table2$IDcolumn1, table1$FKcolumn3 == table3$IDcolumn1, ...}
}
\description{
Get a list of all the Main tables in the database joined with all their relationships with only Include == TRUE columns
}
\examples{
main_joint_tables <- update_db_info(Driver = "{SQL Server};", Database = "Chronogramme", Server = '123.456.7.8', UID = NULL, PWD = NULL, Trusted_Connection = TRUE, Port = 1433,
                                    DIM_Employee$Employee_MainSiteMainObjekt_SiteID == DIM_Site$Site_SiteID,
                                    DIM_Site$Site_RegionID == DIM_Region$Region_ID
                                    )
}
\keyword{Connection}
\keyword{Join}
\keyword{SQL}
\keyword{db}
