% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db_info.R
\name{read_db_fields_csv}
\alias{read_db_fields_csv}
\title{Reads the csv file that is formatted according to the db_fields standard}
\usage{
read_db_fields_csv(csv_path)
}
\arguments{
\item{csv_path}{A String. The path of the db_fields .csv file.}
}
\description{
Returns the db_fields, a DF with columns: KeyType, Table, Column, Type, Comment, RelationshipWithTable, RelationshipWithColumn, Transformation
}
\examples{
db_fields <- read_db_fields_csv("db_fields.csv")
}
\keyword{Read}
\keyword{db_fields}
\keyword{dbfields}
